+++ app/components/Thread/ThreadList.tsx
@@ -4,7 +4,7 @@ import { parseDatumCbor } from '@meshsdk/mesh-csl';
 import { ThreadModal } from './ThreadModal';
 import BlurImage from '../BlurImage';
 import { hexToString } from '../utilities';
-import { BytesField } from './transaction';
+import { createFilterFunctions } from './filters';
 
 interface ThreadListProps {
   network: number | null;
@@ -22,13 +22,25 @@ const ThreadList: React.FC<ThreadListProps> = ({ network, wallet, threads, refre
   const maxPageButtons = 3; // Maximum number of page buttons to show at a time
   const cognoTokenName = sessionStorage.getItem('cognoTokenName');
 
+  // destruc the filter functions
+  const { handleFilterMyThreads, 
+          handleFilterByCategory, 
+          handleFilterMyFriends, 
+          handleFilterTopThreads,
+          handleFilterBySearch
+  } = createFilterFunctions(
+    setFilteredThreads,
+    setCurrentPage,
+    setSearchInput,
+    threads
+  );
+
   useEffect(() => {
     if (selectedThread) {
       document.body.classList.add('overflow-hidden');
     } else {
       document.body.classList.remove('overflow-hidden');
     }
-
     // Cleanup function to remove the class if the component unmounts
     return () => {
       document.body.classList.remove('overflow-hidden');
@@ -43,35 +55,9 @@ const ThreadList: React.FC<ThreadListProps> = ({ network, wallet, threads, refre
       return thread.output.amount.find((asset: Asset) => asset.unit.includes(threadTokenName!));
     });
     setSelectedThread(updatedThread!);
-    // is this needed?
     setCurrentPage(1);
   }, [threads]);
 
-  const handleFilterAll = () => {
-    setFilteredThreads(threads);
-    setCurrentPage(1);
-    setSearchInput('');
-  };
-
-  const handleFilterMyThreads = () => {
-    const filtered = threads.filter(thread => {
-      const parsedDatum = parseDatumCbor(thread.output.plutusData!);
-      return parsedDatum.fields[5].bytes === cognoTokenName;
-    });
-    setFilteredThreads(filtered);
-    setCurrentPage(1);
-    setSearchInput('');
-  };
-
-  const handleFilterByCategory = (category: string) => {
-    const filtered = threads.filter(thread => {
-      const parsedDatum = parseDatumCbor(thread.output.plutusData!);
-      return hexToString(parsedDatum.fields[3].bytes) === category;
-    });
-    setFilteredThreads(filtered);
-    setCurrentPage(1);
-    setSearchInput('');
-  };
 
   const handleThreadClick = (thread: UTxO) => {
     const threadTokenName = thread.output.amount.find((asset: Asset) => asset.unit.includes(process.env.NEXT_PUBLIC_THREAD_MINTER_SCRIPT_HASH!))!.unit.replace(process.env.NEXT_PUBLIC_THREAD_MINTER_SCRIPT_HASH!, '');
@@ -115,30 +101,6 @@ const ThreadList: React.FC<ThreadListProps> = ({ network, wallet, threads, refre
     setCurrentPage(totalPages);
   };
 
-  const handleSearchInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    setSearchInput(value);
-
-    if (value.trim() === '') {
-      setFilteredThreads(threads);
-      setCurrentPage(1);
-      return;
-    }
-
-    const filtered = threads.filter(thread => {
-      const parsedDatum = parseDatumCbor(thread.output.plutusData!);
-      const field0 = hexToString(parsedDatum.fields[0].bytes).toLowerCase();
-      const field1 = hexToString(parsedDatum.fields[1].bytes).toLowerCase();
-      const field5 = parsedDatum.fields[5].bytes.toLowerCase();
-      const searchValue = value.toLowerCase();
-
-      return field0.includes(searchValue) || field1.includes(searchValue) || field5.includes(searchValue);
-    });
-
-    setFilteredThreads(filtered);
-    setCurrentPage(1);
-  };
-
   // Calculate the indices for slicing the filtered threads
   const indexOfLastItem = currentPage * itemsPerPage;
   const indexOfFirstItem = indexOfLastItem - itemsPerPage;
@@ -163,25 +125,14 @@ const ThreadList: React.FC<ThreadListProps> = ({ network, wallet, threads, refre
       {/* Filter Buttons */}
       <div className="flex justify-start my-1">
         <div className="w-1/12"></div> {/* Empty spacer */}
-        <button onClick={handleFilterAll} className="blue-bg blue-bg-hover dark-text py-2 mr-1 rounded w-2/12 font-bold">All Threads</button>
-        <button onClick={handleFilterMyThreads} className="blue-bg blue-bg-hover dark-text py-2 rounded w-2/12 font-bold">My Threads</button>
-        <div className="w-2/12"></div> {/* Empty spacer */}
-        <input
-          placeholder='Search'
-          type="text"
-          value={searchInput}
-          onChange={handleSearchInputChange}
-          className="border rounded dark-text py-2 ml-1 text-center w-4/12"
-          autoComplete="off"
-          data-gramm="false"
-        />
-        <div className="w-1/12"></div> {/* Empty spacer */}
-      </div>
-      <div className="flex justify-start my-1">
-        <div className="w-1/12"></div> {/* Empty spacer */}
-        <button onClick={() => {console.log("Clicked Top Threads")}} className="blue-bg blue-bg-hover dark-text py-1 rounded mr-1 w-2/12">Top Threads</button>
-        <button onClick={() => {console.log("Clicked My Friends")}} className="blue-bg blue-bg-hover dark-text py-1 rounded mr-1 w-2/12">My Friends</button>
-        <select onChange={(e) => { handleFilterByCategory(e.target.value) }} className="blue-bg blue-bg-hover dark-text border py-1 rounded w-2/12 text-center">
+        <select
+          onClick={(e) => {
+            const target = e.target as HTMLSelectElement;
+            handleFilterByCategory(target.value)
+          }}
+          className="blue-bg blue-bg-hover dark-text border py-2 mr-1 rounded w-2/12 text-center font-bold"
+        >
+          <option className="light-bg blue-bg-hover" value="All">All</option>
           <option className="light-bg blue-bg-hover" value="General">General</option>
           <option className="light-bg blue-bg-hover" value="Blockchain">Blockchain</option>
           <option className="light-bg blue-bg-hover" value="News">News</option>
@@ -193,6 +144,23 @@ const ThreadList: React.FC<ThreadListProps> = ({ network, wallet, threads, refre
           <option className="light-bg blue-bg-hover" value="Random">Random</option>
           <option className="light-bg blue-bg-hover" value="Adult">Adult</option>
         </select>
+        <button onClick={handleFilterMyThreads} className="blue-bg blue-bg-hover dark-text py-2 rounded w-2/12 font-bold">My Threads</button>
+        <div className="w-2/12"></div> {/* Empty spacer */}
+        <input
+          placeholder='Search'
+          type="text"
+          value={searchInput}
+          onChange={handleFilterBySearch}
+          className="border rounded dark-text py-2 ml-1 text-center w-4/12"
+          autoComplete="off"
+          data-gramm="false"
+        />
+        <div className="w-1/12"></div> {/* Empty spacer */}
+      </div>
+      <div className="flex justify-start my-1">
+        <div className="w-1/12"></div> {/* Empty spacer */}
+        <button onClick={handleFilterTopThreads} className="blue-bg blue-bg-hover dark-text py-1 rounded mr-1 w-2/12">Top Threads</button>
+        <button onClick={handleFilterMyFriends} className="blue-bg blue-bg-hover dark-text py-1 rounded mr-1 w-2/12">My Friends</button>
         <div className="w-1/12"></div> {/* Empty spacer */}
       </div>